+++ programs/cp-amm/src/instructions/mod.rs
@@ -30,3 +30,6 @@ pub mod ix_close_position;
 pub use ix_close_position::*;
 pub mod ix_split_position;
 pub use ix_split_position::*;
+
+pub mod ix_split_position2;
+pub use ix_split_position2::*;