+++ tests/bankrun-utils/cpAmm.ts
@@ -1780,7 +1780,6 @@ export async function splitPosition(
     reward1Percentage,
   } = params;
   const program = createCpAmmProgram();
-  const poolAuthority = derivePoolAuthority();
   const transaction = await program.methods
     .splitPosition({
       permanentLockedLiquidityPercentage,
@@ -1791,6 +1790,49 @@ export async function splitPosition(
       reward1Percentage,
       padding: new Array(16).fill(0),
     })
+    .accountsPartial({
+      pool,
+      firstPosition,
+      firstPositionNftAccount,
+      secondPosition,
+      secondPositionNftAccount,
+      firstOwner: firstPositionOwner.publicKey,
+      secondOwner: secondPositionOwner.publicKey,
+    })
+    .transaction();
+  transaction.recentBlockhash = (await banksClient.getLatestBlockhash())[0];
+  transaction.sign(firstPositionOwner, secondPositionOwner);
+
+  await processTransactionMaybeThrow(banksClient, transaction);
+}
+
+export type SplitPosition2Params = {
+  firstPositionOwner: Keypair;
+  secondPositionOwner: Keypair;
+  pool: PublicKey;
+  firstPosition: PublicKey;
+  firstPositionNftAccount: PublicKey;
+  secondPosition: PublicKey;
+  secondPositionNftAccount: PublicKey;
+  numerator: number;
+};
+export async function splitPosition2(
+  banksClient: BanksClient,
+  params: SplitPosition2Params
+) {
+  const {
+    pool,
+    firstPositionOwner,
+    secondPositionOwner,
+    firstPosition,
+    secondPosition,
+    firstPositionNftAccount,
+    secondPositionNftAccount,
+    numerator
+  } = params;
+  const program = createCpAmmProgram();
+  const transaction = await program.methods
+    .splitPosition2(numerator)
     .accountsPartial({
       pool,
       firstPosition,